refactor: remove unnecessary types files
