+++ packages/ai/package.json
@@ -31,7 +31,7 @@
     "./tasks/steps/select-image": "./src/tasks/steps/select-image-step.ts",
     "./tasks/steps/visual": "./src/tasks/steps/step-visual.ts",
     "./tasks/steps/visual-image": "./src/tasks/steps/step-visual-image.ts",
-    "./types": "./src/types.ts"
+    "./provider-options": "./src/provider-options.ts"
   },
   "scripts": {
     "typecheck": "tsgo --noEmit"
