+++ packages/ai/src/provider-options.ts
