+++ packages/ai/src/tasks/activities/core/activity-challenge.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import { Output, generateText } from "ai";
 import { z } from "zod";
-import { type ReasoningEffort, buildProviderOptions } from "../../../types";
+import { type ReasoningEffort, buildProviderOptions } from "../../../provider-options";
 import systemPrompt from "./activity-challenge.prompt.md";
 
 const DEFAULT_MODEL = process.env.AI_MODEL_ACTIVITY_CHALLENGE ?? "openai/gpt-5.2";
