+++ packages/ai/src/tasks/activities/core/activity-examples.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import { Output, generateText } from "ai";
 import { z } from "zod";
-import { type ReasoningEffort, buildProviderOptions } from "../../../types";
+import { type ReasoningEffort, buildProviderOptions } from "../../../provider-options";
 import systemPrompt from "./activity-examples.prompt.md";
 
 const DEFAULT_MODEL = process.env.AI_MODEL_ACTIVITY_EXAMPLES ?? "openai/gpt-5.2";
