+++ packages/ai/src/tasks/courses/course-categories.ts
@@ -2,7 +2,7 @@ import "server-only";
 import { COURSE_CATEGORIES } from "@zoonk/utils/categories";
 import { Output, generateText } from "ai";
 import { z } from "zod";
-import { type ReasoningEffort, buildProviderOptions } from "../../types";
+import { type ReasoningEffort, buildProviderOptions } from "../../provider-options";
 import systemPrompt from "./course-categories.prompt.md";
 
 const DEFAULT_MODEL = process.env.AI_MODEL_COURSE_CATEGORIES ?? "google/gemini-3-flash";
