+++ packages/ai/src/tasks/courses/course-description.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import { Output, generateText } from "ai";
 import { z } from "zod";
-import { type ReasoningEffort, buildProviderOptions } from "../../types";
+import { type ReasoningEffort, buildProviderOptions } from "../../provider-options";
 import systemPrompt from "./course-description.prompt.md";
 
 const DEFAULT_MODEL = process.env.AI_MODEL_COURSE_DESCRIPTION || "xai/grok-4-fast-reasoning";
