+++ packages/ai/src/tasks/lessons/lesson-activities.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import { Output, generateText } from "ai";
 import { z } from "zod";
-import { type ReasoningEffort, buildProviderOptions } from "../../types";
+import { type ReasoningEffort, buildProviderOptions } from "../../provider-options";
 import systemPrompt from "./lesson-activities.prompt.md";
 
 const DEFAULT_MODEL = process.env.AI_MODEL_LESSON_ACTIVITIES ?? "openai/gpt-5.2";
