+++ packages/ai/src/tasks/lessons/lesson-kind.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import { Output, generateText } from "ai";
 import { z } from "zod";
-import { type ReasoningEffort, buildProviderOptions } from "../../types";
+import { type ReasoningEffort, buildProviderOptions } from "../../provider-options";
 import systemPrompt from "./lesson-kind.prompt.md";
 
 const DEFAULT_MODEL = process.env.AI_MODEL_LESSON_KIND ?? "google/gemini-2.5-flash-lite";
