--- packages/auth/src/types.ts
@@ -1,2 +1,0 @@
-export type { UserWithRole } from "better-auth/plugins";
-export type { CoursePermission } from "./permissions";
