+++ packages/core/package.json
@@ -38,6 +38,7 @@
     "@vercel/blob": "2.0.1",
     "@zoonk/ai": "workspace:*",
     "@zoonk/auth": "workspace:*",
+    "better-auth": "1.4.17",
     "@zoonk/db": "workspace:*",
     "@zoonk/mailer": "workspace:*",
     "@zoonk/utils": "workspace:*"
