+++ packages/core/src/orgs/org-permissions.ts
@@ -1,6 +1,6 @@
 import "server-only";
 import { auth } from "@zoonk/auth";
-import { type CoursePermission } from "@zoonk/auth/types";
+import { type CoursePermission } from "@zoonk/auth/permissions";
 import { safeAsync } from "@zoonk/utils/error";
 import { headers } from "next/headers";
 import { cache } from "react";
