+++ packages/core/src/types.ts
@@ -2,7 +2,8 @@ import { type auth } from "@zoonk/auth";
 
 export type AuthOrganization = typeof auth.$Infer.Organization;
 
-export type { CoursePermission, UserWithRole } from "@zoonk/auth/types";
+export type { CoursePermission } from "@zoonk/auth/permissions";
+export type { UserWithRole } from "better-auth/plugins";
 
 export type {
   Account,
